+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 New:
 - Compose UI implementation for `Box`.
+- Margin modifier support for HTML DOM layouts.
 
 Changed:
 - Disable klib signature clash checks for JS compilations. These occasionally occur as a result of Compose compiler behavior, and are safe to disable (the first-party JetBrains Compose Gradle plugin also disables them).